Fixed endless loop
A previous update had overwritten the toString() function of classes in
order to provide the name of the class instead of the automatically
generated constructor implementation. That very same change prevented
an internal function from detecting the parameter names of such a
constructor. Now using the correct toString() function for the parameter
names detection.